+++ test/registered/unit/server_args/test_server_args.py
@@ -3418,22 +3418,30 @@ class TestDcpCommBackendDefault(CustomTestCase):
 
     @override_platform(is_cuda=True, is_hip=False)
     def test_fi_a2a_where_supported(self):
-        with patch.object(parallel_hook, "is_fi_a2a_supported", return_value=True):
+        with patch(
+            "sglang.srt.arg_groups.overrides.is_fi_a2a_supported", return_value=True
+        ):
             self.assertEqual(self._resolved(dcp_size=4), "fi_a2a")
 
     @override_platform(is_cuda=True, is_hip=False)
     def test_a2a_on_cuda_without_mnnvl(self):
-        with patch.object(parallel_hook, "is_fi_a2a_supported", return_value=False):
+        with patch(
+            "sglang.srt.arg_groups.overrides.is_fi_a2a_supported", return_value=False
+        ):
             self.assertEqual(self._resolved(dcp_size=4), "a2a")
 
     @override_platform(is_cuda=False, is_hip=False)
     def test_ag_rs_off_cuda(self):
-        with patch.object(parallel_hook, "is_fi_a2a_supported", return_value=False):
+        with patch(
+            "sglang.srt.arg_groups.overrides.is_fi_a2a_supported", return_value=False
+        ):
             self.assertEqual(self._resolved(dcp_size=4), "ag_rs")
 
     @override_platform(is_cuda=True, is_hip=False)
     def test_explicit_value_wins(self):
-        with patch.object(parallel_hook, "is_fi_a2a_supported", return_value=True):
+        with patch(
+            "sglang.srt.arg_groups.overrides.is_fi_a2a_supported", return_value=True
+        ):
             self.assertEqual(
                 self._resolved(dcp_size=4, dcp_comm_backend="ag_rs"), "ag_rs"
             )
